+++ pkg/objstore/cos/cos.go
@@ -333,8 +333,19 @@ func (b *Bucket) getRange(ctx context.Context, name string, off, length int64) (
 		runutil.ExhaustCloseWithLogOnErr(b.logger, resp.Body, "cos get range obj close")
 		return nil, err
 	}
+	// Add size info into reader to pass it to Upload function.
+	r := objectSizerReadCloser{ReadCloser: resp.Body, size: resp.ContentLength}
+	return r, nil
+}
 
-	return resp.Body, nil
+type objectSizerReadCloser struct {
+	io.ReadCloser
+	size int64
+}
+
+// ObjectSize implement objstore.ObjectSizer.
+func (o objectSizerReadCloser) ObjectSize() (int64, error) {
+	return o.size, nil
 }
 
 // Get returns a reader for the given object name.